Initial malloc implementation
Right now it just calls sbrk(), Later we will need to make sure that
this can call the grub equivalent (or create our own).

No metadata is stored with the malloc call so free is effectively a
no-op for now. Yes this means it does leak memory all over the place but
for now it's good enough.